requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
underthesea>=1.3.5
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
nltk>=3.8.0
wordcloud>=1.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pydantic>=2.6.0
psycopg2-binary>=2.9.0
feedparser>=6.0.10
python-dateutil>=2.9.0
sentencepiece>=0.1.99
APScheduler>=3.10.4
